splitted some fucntions

+++ libft/ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: aheinane <aheinane@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/24 15:11:20 by aheinane          #+#    #+#             */
-/*   Updated: 2024/09/26 13:16:53 by aheinane         ###   ########.fr       */
+/*   Updated: 2024/09/26 14:52:44 by aheinane         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str )
+size_t	ft_strlen(const char *str )
 {
 	size_t	i;
 	size_t	count;
